Organize code that handles settings in the modal
